Functional tests passed

+++ responses/thread.go
@@ -6,12 +6,12 @@ import (
 
 type Thread struct {
 	IsNew   bool      `json:"-"`
-	ID      int64    `json:"id"`
+	ID      int64     `json:"id"`
 	Title   string    `json:"title"`
 	Author  string    `json:"author"`
 	Forum   string    `json:"forum"`
 	Message string    `json:"message"`
-	Votes   int32     `json:"votes"`
+	Votes   int64     `json:"votes"`
 	Slug    string    `json:"slug,omitempty"`
 	Created time.Time `json:"created"`
 }
